refactor(allocate): mirror alloc module paths
Module structure now mirrors std::/alloc:: layouts so canonical
import paths line up:

- allocate::boxed::Box                    (was: allocate::Box)
- allocate::vec::Vec                      (was: allocate::Vec)
- allocate::sync::{Arc, Weak}             (was: allocate::{Arc, Weak})
- allocate::collections::HashMap          (was: allocate::HashMap)
- allocate::talc::{TalcAlloc, CacheTalcLock, Manual}
                                          (was: allocate::{TalcAlloc, …})

The Allocator trait + Global / AllocError / Layout / allocate /
allocate_zeroed / deallocate helpers stay at the crate root (no
natural submodule home).

Workspace consumers updated to the canonical paths via mechanical
import rewrites. Tests + clippy + smoke + check-no-api2 all clean.
