+++ rust/allocate/src/collections/hashmap.rs
