+++ rust/javm-bench/examples/publish_profile.rs
@@ -22,7 +22,7 @@ use std::time::Instant;
 #[cfg(all(target_os = "linux", target_arch = "x86_64"))]
 use allocate::Global;
 #[cfg(all(target_os = "linux", target_arch = "x86_64"))]
-use allocate::Vec as AVec;
+use allocate::vec::Vec as AVec;
 #[cfg(all(target_os = "linux", target_arch = "x86_64"))]
 use javm_cap::cap::Cap;
 #[cfg(all(target_os = "linux", target_arch = "x86_64"))]
