+++ rust/nub-arch-x86/src/call_loop.rs
@@ -94,7 +94,7 @@ extern crate alloc;
 use alloc::vec;
 use alloc::vec::Vec;
 
-use allocate::TalcAlloc;
+use allocate::talc::TalcAlloc;
 use javm_cap::cap::{Cap, CapHashOrRef};
 use javm_cap::hash::{Blake2b256, Hash};
 use javm_cap::slot::SlotIdx;
@@ -783,7 +783,7 @@ fn build_transient_instance_cap(
     image_hash: CapHash,
     image_hash_chain: CapHash,
 ) -> Cap<TalcAlloc> {
-    use allocate::Vec as AVec;
+    use allocate::vec::Vec as AVec;
     let alloc = cache.allocator();
     Cap::Instance(javm_cap::instance::InstanceCap {
         image_hash_chain,
