+++ rust/ssz/src/lib.rs
@@ -26,7 +26,8 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 extern crate alloc;
 
-use allocate::{Allocator, Global, Vec};
+use allocate::vec::Vec;
+use allocate::{Allocator, Global};
 use digest::Digest;
 use digest::typenum::U32;
 
@@ -105,7 +106,7 @@ pub trait Encode {
     ///
     /// This is provided for convenience in callers that don't care about
     /// allocator threading. The default impl encodes into an
-    /// `allocate::Vec<u8, Global>` and copies into a standard
+    /// `allocate::vec::Vec<u8, Global>` and copies into a standard
     /// `alloc::vec::Vec<u8>`. Implementations may override for efficiency.
     fn as_ssz_bytes(&self) -> alloc::vec::Vec<u8> {
         let mut out = alloc::vec::Vec::with_capacity(self.ssz_bytes_len());
